Add types to UserContext
